+++ t/unit/test_canvas.py
@@ -1,7 +1,5 @@
 import uuid
 
-from celery import group
-
 
 class test_Canvas:
 
@@ -33,16 +31,3 @@ class test_Canvas:
         assert t_reply_to_app == t_reply_to_opt
         # reply_to must be thread-relative.
         assert t_reply_to_opt != s.options['reply_to']
-
-    def test_group_unroll(self, subtests):
-        @self.app.task
-        def test_task(a, b):
-            return
-
-        with subtests.test("single item"):
-            c = group(test_task.s(1, 2)) | test_task.s(1)
-            assert str(c) == "t.unit.test_canvas.test_task(1, 2) | test_task(1)"
-
-        with subtests.test("regen"):
-            c = group(test_task.s(1, 2) for _ in range(1)) | test_task.s(1)
-            assert str(c) == "t.unit.test_canvas.test_task(1, 2) | test_task(1)"
